@@ -19,8 +19,8 @@ android {
     }
     namespace 'com.example.inventory'
     compileOptions {
-        targetCompatibility JavaVersion.VERSION_17
-        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_19
+        sourceCompatibility JavaVersion.VERSION_19
     }
     buildToolsVersion '32.0.0'
 }
