@@ -1,30 +1,40 @@
 package com.example.inventory;
 
 import android.content.Context;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
-import android.widget.SpinnerAdapter;
+import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 
-import com.example.inventory.data.DbHelper;
-
-import java.util.Map;
+import java.util.ArrayList;
 
 
-public class InventoryArrayAdapter extends ArrayAdapter<InventoryItem> implements SpinnerAdapter {
+public class InventoryArrayAdapter extends ArrayAdapter<InventoryItem>{
 
-    Map<String, InventoryItem> values;
-    DbHelper database;
+    public InventoryArrayAdapter(@NonNull Context context, ArrayList<InventoryItem> inventory) {
+        super(context, 0, inventory);
+    }
 
-    public InventoryArrayAdapter(@NonNull Context context, int resource) {
-        super(context, android.R.layout.simple_spinner_item);
-        this.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-        database = new DbHelper( context );
-        values = database.getInventoryItemWithQuantities("1");
-        for (Map.Entry<String, InventoryItem> value: values.entrySet()
-             ) {
-                this.add(value.getValue());
+    public View getView(int position,
+                        View convertView,
+                        ViewGroup parent){
+
+        InventoryItem item = getItem(position);
+        if (convertView == null) {
+            convertView = LayoutInflater.from(getContext()).inflate(R.layout.inventory_list_item,
+                    parent, false);
         }
+        TextView nameView = (TextView) convertView.findViewById(R.id.name);
+        TextView quantityView = (TextView) convertView.findViewById(R.id.quantity);
+
+        nameView.setText(item.itemName);
+        quantityView.setText(Integer.toString(item.itemQuantity));
+
+        return convertView;
+
     }
 
 }
