@@ -1,4 +1,14 @@
 package com.example.inventory;
 
-public record InventoryItem(String name, Integer currentQuantity) {
+
+public class InventoryItem{
+
+    public final String itemName;
+    public final Integer itemQuantity;
+
+    public InventoryItem(String itemName, Integer itemQuantity){
+        this.itemName = itemName;
+        this.itemQuantity = itemQuantity;
+
+    }
 }
